feat(bevy_kbve_net): add npcdb feature flag for shared CreatureRegistry
Move CreatureRegistry, CreatureConfig, slot helpers, and
build_creature_registry() into bevy_kbve_net behind an optional npcdb
feature. Both isometric-game and axum-kbve enable the feature so they
share identical creature definitions.

Wire process_creature_captures on the server to validate capture
requests against the registry pool_size before broadcasting.
