+++ apps/kbve/axum-kbve/Cargo.toml
@@ -59,7 +59,7 @@ lightyear = { version = "0.26", default-features = false, features = [
     "raw_connection",
 ] }
 lightyear_avian3d = { version = "0.26", default-features = false, features = ["3d"] }
-bevy_kbve_net = { path = "../../../packages/rust/bevy/bevy_kbve_net" }
+bevy_kbve_net = { path = "../../../packages/rust/bevy/bevy_kbve_net", features = ["npcdb"] }
 futures-util = "0.3"
 
 # Workspace path dependencies
