+++ apps/kbve/isometric/src-tauri/Cargo.toml
@@ -50,8 +50,7 @@ serde_json = "1"
 avian3d = { version = "0.5", default-features = false, features = ["3d", "f32", "parry-f32", "debug-plugin"] }
 bevy_inventory = { path = "../../../../packages/rust/bevy/bevy_inventory" }
 bevy_cam = { path = "../../../../packages/rust/bevy/bevy_cam" }
-bevy_kbve_net = { path = "../../../../packages/rust/bevy/bevy_kbve_net" }
-bevy_npc = { path = "../../../../packages/rust/bevy/bevy_npc" }
+bevy_kbve_net = { path = "../../../../packages/rust/bevy/bevy_kbve_net", features = ["npcdb"] }
 lightyear = { version = "0.26", default-features = false, features = [
     "client",
     "replication",
