+++ packages/rust/bevy/bevy_kbve_net/Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2021"
 description = "Shared lightyear protocol for KBVE multiplayer — replicated components, inputs, channels"
 
+[features]
+default = []
+npcdb = ["bevy_npc"]
+
 [dependencies]
 bevy = { version = "0.18", default-features = false, features = ["bevy_state", "bevy_render"] }
 lightyear = { version = "0.26", default-features = false, features = [
@@ -14,3 +18,4 @@ lightyear = { version = "0.26", default-features = false, features = [
 ] }
 avian3d = { version = "0.5", default-features = false, features = ["3d", "f32", "parry-f32", "serialize"] }
 serde = { version = "1", features = ["derive"] }
+bevy_npc = { path = "../bevy_npc", optional = true }
